+++ protocol/src/protocol/model/pFood.java
@@ -4,18 +4,18 @@ package protocol.model;
  * @author apomosov
  */
 public final class pFood{
-  private int x;
-  private int y;
-  public pFood(int x, int y){
+  private double x;
+  private double y;
+  public pFood(double x, double y){
     this.x=x;
     this.y=y;
   }
 
-  public int getX() {
+  public double getX() {
     return x;
   }
 
-  public int getY() {
+  public double getY() {
     return y;
   }
 }
