+++ protocol/src/protocol/model/pVirus.java
@@ -4,18 +4,18 @@ package protocol.model;
  * Created by Max on 15.12.2016.
  */
 public class pVirus {
-    private int x;
-    private int y;
-    public pVirus(int x, int y){
+    private double x;
+    private double y;
+    public pVirus(double x, double y){
         this.x=x;
         this.y=y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
